pyproject.toml
[project]
name = "playground-1-sentencetransformers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.3",
]
